tempPipeline10-27-811/requirements.txt
# Complete Pipeline Dependencies
# tempPipeline10-27-811

# Core Python
python>=3.9

# Database (FDC/PostgreSQL)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Configuration & Environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Vision Model (OpenAI GPT-4V/GPT-5)
openai>=1.0.0

# Data Validation
pydantic>=2.0.0

# Testing & Evaluation
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0  # Progress bars for batch processing
